added a "zuckwatch" cog and transfered the zuckpass command
