+++ cogs/general.py
@@ -26,14 +26,5 @@ class General(commands.Cog):
         print(f"Interrogative: {arg}")
         await ctx.send(input("Response: "))
 
-    @commands.command()
-    async def zuckpass(self, ctx, *, arg):
-        """| Checks a single password against the Zuckwatch API"""
-        response = requests.post('https://k1ozwahixa.execute-api.us-east-1.amazonaws.com/dev/password',json={'password': arg})
-        if response.status_code != 400:
-            await ctx.send("Your submitted password is correct.")
-        else:
-            await ctx.send("Your submitted password is incorrect.")
-
 def setup(bot):
     bot.add_cog(General(bot))
